+++ src/main/java/coordinate/domain/figure/Rectangle.java
@@ -1,4 +1,6 @@
-package coordinate.domain;
+package coordinate.domain.figure;
+
+import coordinate.domain.point.PointGroup;
 
 public class Rectangle extends AbstractFigure {
     static final int SIZE_OF_POINTS = 4;
