+++ src/main/java/coordinate/domain/point/Coordinate.java
@@ -1,4 +1,4 @@
-package coordinate.domain;
+package coordinate.domain.point;
 
 import java.util.HashMap;
 import java.util.Map;
