+++ src/main/java/coordinate/domain/point/PointGroupFactory.java
@@ -1,4 +1,4 @@
-package coordinate.domain;
+package coordinate.domain.point;
 
 import java.util.List;
 import java.util.regex.Pattern;
