+++ src/test/java/coordinate/domain/point/PointTest.java
@@ -1,4 +1,4 @@
-package coordinate.domain;
+package coordinate.domain.point;
 
 import org.junit.jupiter.api.Test;
